refactor: move diracx function in Framework
